+++ x/observer/keeper/msg_server_update_operational_flags_test.go
@@ -2,10 +2,12 @@ package keeper_test
 
 import (
 	"testing"
+	"time"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/stretchr/testify/require"
 
+	"github.com/zeta-chain/node/pkg/ptr"
 	keepertest "github.com/zeta-chain/node/testutil/keeper"
 	"github.com/zeta-chain/node/testutil/sample"
 	authoritytypes "github.com/zeta-chain/node/x/authority/types"
@@ -26,10 +28,12 @@ func TestMsgServer_UpdateOperationalFlags(t *testing.T) {
 
 		// test initial set
 		restartHeight := int64(100)
+		signerBlockTimeOffset := ptr.Ptr(time.Second)
 		msg := types.MsgUpdateOperationalFlags{
 			Creator: admin,
 			OperationalFlags: types.OperationalFlags{
-				RestartHeight: restartHeight,
+				RestartHeight:         restartHeight,
+				SignerBlockTimeOffset: signerBlockTimeOffset,
 			},
 		}
 		keepertest.MockCheckAuthorization(&authorityMock.Mock, &msg, nil)
@@ -39,13 +43,16 @@ func TestMsgServer_UpdateOperationalFlags(t *testing.T) {
 		operationalFlags, found := k.GetOperationalFlags(ctx)
 		require.True(t, found)
 		require.Equal(t, restartHeight, operationalFlags.RestartHeight)
+		require.Equal(t, signerBlockTimeOffset, operationalFlags.SignerBlockTimeOffset)
 
 		// verify that we can set it again
 		restartHeight = 101
+		signerBlockTimeOffset = ptr.Ptr(time.Second * 2)
 		msg = types.MsgUpdateOperationalFlags{
 			Creator: admin,
 			OperationalFlags: types.OperationalFlags{
-				RestartHeight: restartHeight,
+				RestartHeight:         restartHeight,
+				SignerBlockTimeOffset: signerBlockTimeOffset,
 			},
 		}
 		keepertest.MockCheckAuthorization(&authorityMock.Mock, &msg, nil)
@@ -55,6 +62,7 @@ func TestMsgServer_UpdateOperationalFlags(t *testing.T) {
 		operationalFlags, found = k.GetOperationalFlags(ctx)
 		require.True(t, found)
 		require.Equal(t, restartHeight, operationalFlags.RestartHeight)
+		require.Equal(t, signerBlockTimeOffset, operationalFlags.SignerBlockTimeOffset)
 	})
 
 	t.Run("cannot update operational flags if not authorized", func(t *testing.T) {